lzma/zstd: port state machine, bit model, ZSTD constants + spec docs
Three new algorithmic modules ported from omnizip Ruby:

LZMA state machine (omnizip-lzma/src/state.rs):
  - 12-state machine with const-array transition tables
  - on_literal/on_match/on_rep/on_short_rep methods
  - Transition tables match XZ Utils reference exactly:
    LIT_STATES     = [0,0,0,0,1,2,3,4,5,6,4,5]
    MATCH_STATES   = [7,7,7,7,7,7,7,10,10,10,10,10]
    REP_STATES     = [8,8,8,8,8,8,8,11,11,11,11,11]
    SHORT_REP_STATES = [9,9,9,9,9,9,9,11,11,11,11,11]
  - 9 tests (transition tables verified, context queries, sequences)

LZMA bit model (omnizip-lzma/src/bit_model.rs):
  - BitModel: adaptive u16 probability in [1, TOTAL-1]
  - BitModelArray: Vec<BitModel> for probability tables
  - Update formula matches XZ Utils RC_BIT_* macros:
    bit 0: prob += (TOTAL - prob) >> MOVE_BITS
    bit 1: prob -= prob >> MOVE_BITS
  - Constants retyped: BIT_MODEL_TOTAL u32->u16, MOVE_BITS u32->u8
  - 8 tests (adaptation, convergence, array operations)

ZSTD constants (omnizip-zstd/src/constants.rs):
  - Frame magic, block types, literals modes, FSE accuracy logs
  - Window log range, Huffman limits, compression level range
  - All values from RFC 8878 §3-4
  - 7 tests (magic verification, level range, FSE accuracy)

Three-tier spec documentation:
  TODO.spec/ — normative bit-level specs (5 files):
    01-lzma-overview, 02-lzma-range-coder, 03-lzma-state-machine,
    10-zstd-frame, 14-zstd-fse
  TODO.requirements/ — acceptance criteria (3 files):
    README with 9 hard requirements, LZMA requirements, ZSTD requirements
  TODO.references/ — source mappings (3 files):
    README with source hierarchy, LZMA source map (31 Ruby files),
    ZSTD source map (14 Ruby files)

66 workspace tests pass (was 42). Clippy clean. fmt clean.

+++ omnizip-lzma/src/constants.rs
@@ -25,15 +25,15 @@
 pub const TOP: u32 = 0x0100_0000;
 
 /// Total probability range for bit models (`2^11`).
-pub const BIT_MODEL_TOTAL: u32 = 0x800;
+pub const BIT_MODEL_TOTAL: u16 = 0x800;
 
 /// Number of bits to shift on each probability update. Smaller = faster
 /// adaptation; larger = more stable.
-pub const MOVE_BITS: u32 = 5;
+pub const MOVE_BITS: u8 = 5;
 
 /// Initial probability value for a fresh bit model (`BIT_MODEL_TOTAL / 2`,
 /// i.e. 50% probability).
-pub const INIT_PROBS: u32 = BIT_MODEL_TOTAL >> 1;
+pub const INIT_PROBS: u16 = BIT_MODEL_TOTAL >> 1;
 
 /// Number of bits used in direct-bit encoding.
 pub const NUM_DIRECT_BITS: u32 = 8;
